if defauldelivery do not install an additional LDA on user edit
If DEFAULT_DELIVERY is defined in vpopmail, do not allow user's .qmail modifications. In fact, when an unknown (for qmailadmin) delivery agent is already defined there, and qmailadmin adds another one, users will get two messages for each delivery in their mailbox.
Of course in this case the autoresponder is out of question, but users can create their out of office message by means of a plugin installed in their webmail.